Created the "models" folder for unique classes and moved calendar.dart and user.dart to models
