+++ lib/backend/models/user.dart
