+++ lib/backend/type_driver.dart
@@ -1,7 +1,7 @@
-enum TypeDriver{
- None,
- A,
- B,
- AB,
- Nao_Habilitado,
+enum TypeDriver {
+  None,
+  A,
+  B,
+  AB,
+  Nao_Habilitado,
 }
